backend/pyproject.toml
[project]
name = "outfit-recommender-backend"
version = "0.1.0"
description = "Multi-agent outfit recommendation system backend"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "langgraph>=0.0.1",
    "langchain-google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.0.0",
    "google-generativeai>=0.3.0",
    "Pillow>=10.0.0",
    "dspy-ai>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pyright]
typeCheckingMode = "basic"
pythonVersion = "3.11"

[dependency-groups]
dev = [
    "grandalf>=0.8",
]
